Moons are textured
The moons have been changed to .OBJ format, and they also have a texture
applied.

Most changes are in Moonmodel.h and moonmodel.cpp, with the new
information added to the .scene file, and with one Init() call added to
the world.cpp file.

Let me know ASAP if something doesn't work.